requirements.txt
tensorflow==2.14.0
numpy==1.25.2
Pillow==10.1.0
pyttsx3==2.90
matplotlib==3.8.0
